Create Retail_Guru.R
To find the common monthly basket across customers. (Hint: is this a basket of specific items
or a basket of groups?). Monthly baskets are the one time large baskets that people tend to
buy at the start of the month(but not strictly in that period) and to find the items that can be sold in combo and the reason for these items to coexist in thecombo.